✨(video-clubs) add create and delete functionality
Added a new video club directory with a create and delete functionality. The create functionality allows the user to create a new video club and the delete functionality allows the user to delete a video club.
Also changed a bit the way the video clubs are fetched from the database. Now the video clubs are fetched.
Small changes in the sidebar layout.